+++ CHANGELOG.rst
@@ -5,6 +5,19 @@ dev
 ---
 Features
 ++++++++
+* mock simulation
+* consistent model power to mock to field power
+* explicit dependency checks and cache
+
+Enhancements
+++++++++++++
+* gridding now part of PS class
+* consistent tests for sky map to power spectrum
+
+v0.4.0
+------
+Features
+++++++++
 * a base class for better structure
 * model power spectrum
 * power spectrum weights and convolution